requirements.txt
apturl==0.5.2
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.2
certifi==2020.6.20
cffi==1.15.0
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==36.0.2
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.5.0
distro-info==1.0
dnspython==2.1.0
duplicity==0.8.20
email-validator==1.1.3
fasteners==0.14.1
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
future==0.18.2
greenlet==1.1.2
holidays==0.49
httplib2==0.18.1
idna==3.3
importlib-metadata==4.0.1
itsdangerous==2.0.1
jeepney==0.7.1
Jinja2==3.0.3
keyring==23.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.5
lockfile==0.12.2
lorem-text==2.1
louis==3.18.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.5
more-itertools==4.2.0
mysql-connector==2.2.9
mysql-connector-python==8.0.28
netifaces==0.10.9
oauthlib==3.1.0
olefile==0.46
paramiko==2.7.2
pexpect==4.8.0
Pillow==8.4.0
protobuf==3.19.4
pycairo==1.16.2
pycparser==2.21
pycups==2.0.1
PyGObject==3.40.1
PyJWT==1.7.1
pymacaroons==0.13.0
PyMySQL==1.0.2
PyNaCl==1.4.0
pyRFC3339==1.1
python-apt==2.2.1
python-dateutil==2.8.1
python-debian===0.1.39ubuntu1
python-dotenv==0.19.2
pytz==2021.1
pyxdg==0.27
PyYAML==5.3.1
reportlab==3.5.66
requests==2.25.1
SecretStorage==3.3.1
simplejson==3.17.2
six==1.16.0
SQLAlchemy==1.4.28
systemd-python==234
ubuntu-advantage-tools==27.4
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.5
Werkzeug==2.0.2
WTForms==3.0.0
xdg==5
xkit==0.0.0
zipp==1.0.0
